third_party/WebKit/LayoutTests/crypto/importKey-expected.txt

Tests cypto.subtle.importKey.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'HMAC'
PASS key.algorithm.hash.name is 'SHA-256'
PASS key.usages.join(',') is 'encrypt,sign'
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'AES-CBC'
PASS key.usages.join(',') is ''
PASS key.type is 'secret'
PASS key.extractable is false
PASS key.algorithm.name is 'AES-CBC'
PASS key.usages.join(',') is ''
PASS key.type is 'secret'
PASS key.extractable is false
PASS key.algorithm.name is 'AES-CBC'
PASS key.usages.join(',') is ''
rejected with null
PASS crypto.subtle.importKey('invalid format', data, algorithm, extractable, keyUsages) threw exception TypeError: Invalid keyFormat argument.
PASS crypto.subtle.importKey(keyFormat, data, algorithm, extractable, ['SIGN']) threw exception TypeError: Invalid keyUsages argument.
PASS crypto.subtle.importKey('invalid format', data, algorithm, extractable, ['SIGN']) threw exception TypeError: Invalid keyFormat argument.
PASS crypto.subtle.importKey(keyFormat, [], algorithm, extractable, keyUsages) threw exception TypeError: Invalid keyData argument.
PASS crypto.subtle.importKey(keyFormat, null, algorithm, extractable, keyUsages) threw exception TypeError: Invalid keyData argument.
PASS crypto.subtle.importKey(keyFormat, data, {name: 'hmac'}, extractable, keyUsages) threw exception TypeError: Algorithm: HMAC: HmacParams: hash: Missing or not a dictionary.
PASS crypto.subtle.importKey(keyFormat, data, {name: 'sha-1'}, extractable, keyUsages) threw exception NotSupportedError: Algorithm: SHA-1: Unsupported operation.
PASS successfullyParsed is true

TEST COMPLETE
